Encapsulate functionality into supplementary_plot_helpers.py
